Initial implementation of subgroup search - needs debugging
Implemented an algorithm to find the subgroup of group elements
satisfying a given property. It currently works in most cases,
however there are some exceptions and it needs debugging.

Wrote an utility function _insert_point_in_base which takes
a BSGS and related structures (transversals, orbits) and
inserts a given point in a given position in the base. The
changed base/strong generating set are guaranteed to be valid
only partially, up to the position at which the new point was
inserted.

Wrote some tests for the new functions introduced.