First steps in improving disjoint cycle notation.
The goal is to be able to specify the disjoint cycle representation
of a permutation without listing the singleton cycles.

This commit removes the check for all numbers from 0 through n - 1
being present in the permutation; adds the attribute _size for
a permutation, and when a new permutation is created in cyclic form
sets _size to the maximum point moved. Then, multiplication of
permutations in cyclic from is handled by extending the permutation
of lesser size so that it matches the size of the other permutation.